backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
python-dotenv==1.0.1
openpyxl==3.1.5
django-cors-headers



